+++ databento/live/__init__.py
@@ -1,10 +1,11 @@
-from typing import Union
+from typing import Callable, Union
 
 import databento_dbn
 
 
 AUTH_TIMEOUT_SECONDS: float = 2
 CONNECT_TIMEOUT_SECONDS: float = 5
+
 
 DBNRecord = Union[
     databento_dbn.MBOMsg,
@@ -19,3 +20,6 @@ DBNRecord = Union[
     databento_dbn.SystemMsg,
     databento_dbn.ErrorMsg,
 ]
+
+RecordCallback = Callable[[DBNRecord], None]
+ExceptionCallback = Callable[[Exception], None]