+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Enhancements
 - Added `symbology_map` property to `Live` client
+- Changed `Live.add_callback` and `Live.add_stream` to accept an exception callback
+- Changed `Live.add_callback` and `Live.add_stream` `func` parameter to `record_callback`
 
 ## 0.14.1 - 2023-06-16